+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/container/MuxerContainer.kt
@@ -1,4 +1,4 @@
-package com.llfbandit.record.record.container;
+package com.llfbandit.record.record.container
 
 import android.media.MediaCodec
 import android.media.MediaFormat