+++ record_web/lib/encoder/encoder.dart
@@ -1,0 +1,11 @@
+import 'dart:typed_data';
+
+import 'package:record_web/js/js_interop/core.dart';
+
+abstract class Encoder {
+  void encode(Int16List buffer);
+
+  Blob finish();
+
+  void cleanup();
+}