+++ record_web/lib/webm_duration_fix.dart
@@ -1,7 +1,0 @@
-import 'dart:html';
-import 'dart:js';
-
-import 'package:js/js.dart';
-
-@JS('jsFixWebmDuration')
-external dynamic fixWebmDuration(Blob blob, num duration, JsObject options);