+++ record_windows/windows/record_config.h
@@ -4,9 +4,22 @@
 
 namespace record_windows
 {
+
+	struct AudioEncoder
+	{
+		const std::string aacLc = std::string("aacLc");
+		const std::string aacEld = std::string("aacEld");
+		const std::string aacHe = std::string("aacHe");
+		const std::string amrNb = std::string("amrNb");
+		const std::string amrWb = std::string("amrWb");
+		const std::string opus = std::string("opus");
+		const std::string flac = std::string("flac");
+		const std::string pcm16bits = std::string("pcm16bits");
+	};
+
 	struct RecordConfig
 	{
-		std::string encoderName = "aacLc";
+		std::string encoderName = AudioEncoder().aacLc;
 		std::string deviceId = NULL;
 		int bitRate = 128000;
 		int sampleRate = 44100;